[FIX] mass_mailing_custom_unsubscribe: Allow redirect to mailing list manager on unsubscribe
When unsubscribing from a list, you cannot select a reason until you go to the mailing list manager.
Now, you can choose with a system parameter to redirect directly to the mailing list manager and give the user in the first place an option to select a reason to opt out.